draft accepted by CSAE 2013, revision required.
Review Result from the committee:
The paper proposes the research on high-level abstraction for data-flow drived concurrent
programming using the functional programming language Haskell. The topic is interesting
and has a practical value. But there are some shortcomings in this paper. (1) It's better
to give the simulated experimental results, compare the results of proposed method with
others and extend the analysis about the results. (2) Please double check the English
expression carefully, especially in the abstract.